rename(place) : 관리자 페이지 Admin네이밍 추가
