+++ src/main/java/kr/co/mapspring/place/service/AdminPlaceService.java
@@ -1,4 +1,4 @@
-package kr.co.mapspring.place.service.impl;
+package kr.co.mapspring.place.service;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -7,29 +7,27 @@ import kr.co.mapspring.global.exception.place.PlaceAlreadyExistsException;
 import kr.co.mapspring.global.exception.place.PlaceNotFoundException;
 import kr.co.mapspring.global.exception.place.RegionNotFoundException;
 import kr.co.mapspring.global.exception.place.ScenarioNotFoundException;
-import kr.co.mapspring.place.dto.ReadPlaceDto;
-import kr.co.mapspring.place.dto.SavePlaceDto;
+import kr.co.mapspring.place.dto.AdminReadPlaceDto;
+import kr.co.mapspring.place.dto.AdminSavePlaceDto;
 import kr.co.mapspring.place.entity.Place;
 import kr.co.mapspring.place.entity.Region;
 import kr.co.mapspring.place.entity.Scenario;
 import kr.co.mapspring.place.repository.PlaceRepository;
 import kr.co.mapspring.place.repository.RegionRepository;
 import kr.co.mapspring.place.repository.ScenarioRepository;
-import kr.co.mapspring.place.service.PlaceService;
 import lombok.RequiredArgsConstructor;
 
 @Service
 @RequiredArgsConstructor
-public class PlaceServiceImpl implements PlaceService {
+public class AdminPlaceService {
 
 	private final PlaceRepository placeRepository;
 	private final RegionRepository regionRepository;
 	private final ScenarioRepository scenarioRepository;
 
 	// 장소 저장
-	@Override
 	@Transactional
-	public void savePlace(SavePlaceDto.RequestSave request) {
+	public void savePlace(AdminSavePlaceDto.RequestSave request) {
 		
 		Long regionId = request.getRegionId();
 		Long scenarioId = request.getScenarioId();
@@ -61,16 +59,15 @@ public class PlaceServiceImpl implements PlaceService {
 	}
 	
 	// 장소 조회
-	@Override
 	@Transactional(readOnly = true)
-	public ReadPlaceDto.ResponseRead readPlace(ReadPlaceDto.RequestRead request) {
+	public AdminReadPlaceDto.ResponseRead readPlace(AdminReadPlaceDto.RequestRead request) {
 		
 		Long placeId = request.getPlaceId();
 		
 		Place place = placeRepository.findById(placeId)
 				.orElseThrow(PlaceNotFoundException::new);
 		
-		ReadPlaceDto.ResponseRead response = ReadPlaceDto.ResponseRead.from(place);
+		AdminReadPlaceDto.ResponseRead response = AdminReadPlaceDto.ResponseRead.from(place);
 		
 		return response;
 	}
